22-03-14 23:24

+++ Kim/_220215/Boj_9239_스티브잡송.java
@@ -6,7 +6,7 @@ public class Boj_9239_스티브잡송 {
 
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int x = sc.nextInt();
+		double x = sc.nextDouble();
 		//나눗셈......?
 		//접근법조차 전혀 모르겠음 ㅋㅋㅋ ㅠㅠ
 		//자릿수 따져서 10의 배수로 나눠서 몫을 일의 자리에 붙이는 식으로 해야 하나?
